feat(models-backend): close the loop with real transport, vault, and async sync
Implement the models backend closure plan: replace stub/noop paths with
real HTTP transport, credential vault integration, connection identity
guards, and async sync orchestration.

Backend:
- Add HttpModelTransport (reqwest-based) and ModelTransport trait
- Add ConnectionIdentity guard to prevent mid-flight config corruption
- Add test_connection() and sync_models() async methods to ModelService
- Add connection_identity_changed() to ProviderService; reset sync
  status when connection fields change
- Register test_provider_connection and sync_provider_models Tauri
  commands
- Add ConnectionTestResult and SyncModelsResult domain types
- Wire vault and transport into AppState construction

Frontend:
- Add Test Connection and Get Models buttons to ProviderEditor
- Add IPC client bindings for testProviderConnection and
  syncProviderModels
- Show sync status, timestamps, and error states in the UI

Testing:
- Add TestModelTransport (queued responses) for deterministic tests
- Cover connection identity mismatch, credential errors, transport
  errors, and successful sync paths
- Update credential coordinator tests for stricter operation cleanup
  assertions

Other:
- Update implementation plan document with completed status
- Add src-tauri/gen to .prettierignore
- Add reqwest and tokio crate dependencies

+++ src-tauri/src/windows/main.rs
@@ -62,10 +62,10 @@ fn is_geometry_usable<R: Runtime>(app: &tauri::AppHandle<R>, geometry: &WindowGe
 		let scale = monitor.scale_factor();
 		let pos = monitor.position();
 		let size = monitor.size();
-		let mx = pos.x as f64 / scale as f64;
-		let my = pos.y as f64 / scale as f64;
-		let mw = size.width as f64 / scale as f64;
-		let mh = size.height as f64 / scale as f64;
+		let mx = pos.x as f64 / scale;
+		let my = pos.y as f64 / scale;
+		let mw = size.width as f64 / scale;
+		let mh = size.height as f64 / scale;
 		let intersects = geometry.x < mx + mw
 			&& geometry.x + geometry.width > mx
 			&& geometry.y < my + mh
